working bt module
